+++ ExAnoPassado/total_risco.h
@@ -1,4 +1,4 @@
 #ifndef HEADER1_H
 #define HEADER1_H
-long long total_risco(int *vec, int num); 
+long total_risco(int *vec, int num); 
 #endif
